chore: refactor input import
