+++ components/custom/index.ts
@@ -1,1 +1,2 @@
 export { default as PageTitle } from "./PageTitle";
+export { default as Input } from "./Input";
